chore(tests): refactor tests to use common cleanup
