+++ playwright.config.ts
@@ -12,13 +12,10 @@ import { defineConfig, devices } from "@playwright/test";
 
 export default defineConfig({
     testDir: "./tests",
-    /* Run tests in files in parallel */
-    fullyParallel: true,
     /* Fail the build on CI if you accidentally left test.only in the source code. */
     forbidOnly: !!process.env["CI"],
-    /* Retry on CI only */
     retries: 0,
-    /* Opt out of parallel tests on CI. */
+    /* Opt out of parallel tests */
     workers: 1,
     /* Reporter to use. See https://playwright.dev/docs/test-reporters */
     reporter: [["list", { printSteps: true }], ["html"]],
