+++ tests/find-the-expert-tests.spec.ts
@@ -2,20 +2,10 @@ import { expect, test, type Page } from "@playwright/test";
 import { type ChildProcess } from "child_process";
 
 import { TestSetup } from "~/tests/helpers/test-setup";
-import treeKill from "tree-kill";
-
-import { promisify } from "node:util";
 import { FindTheExpertPage } from "~/tests/find-the-expert-page";
 import { DbHelper } from "~/tests/helpers/db";
 import { USER_EMAIL, USER_NAME, UserDbHelper } from "~/tests/helpers/db/user";
 import { SurveyDbHelper } from "~/tests/helpers/db/survey";
-
-const treeKillAsPromised = promisify(treeKill);
-const killAllProcesses = async (process: ChildProcess) => {
-    if (process?.pid) {
-        await treeKillAsPromised(process.pid);
-    }
-};
 
 let nextProcess: ChildProcess;
 let findTheExpertPage: FindTheExpertPage;
@@ -23,12 +13,14 @@ let dbHelper: DbHelper;
 let userDbHelper: UserDbHelper;
 let surveyDbHelper: SurveyDbHelper;
 let page: Page;
+let cleanup: () => Promise<void>;
 
 test.beforeAll(async ({ browser }) => {
     const setup = await TestSetup.setup(browser, true);
     page = setup.page;
     dbHelper = setup.dbHelper;
     nextProcess = setup.nextProcess;
+    cleanup = setup.cleanup;
 
     userDbHelper = new UserDbHelper(dbHelper.getClient());
     surveyDbHelper = new SurveyDbHelper(dbHelper.getClient());
@@ -36,9 +28,7 @@ test.beforeAll(async ({ browser }) => {
 });
 
 test.afterAll(async () => {
-    await dbHelper.getContainer().stop();
-    await page.close();
-    await killAllProcesses(nextProcess);
+    await cleanup();
 });
 
 test.describe("Desktop tests", () => {
